Add extras as `pinecone-with-*`
This avoids problems where the previous `run_constrainted`
approach was leading to downstream environments that are unsolvable.

While it isn't great that the dependencies aren't constrained
when they *aren't* installed with an extra, this is the same
behavior as on PyPI so should be what the developers of both
`pinecone` and packages that have `pinecone` as a dependency
are expecting.